Update patch set 1
Patch Set 1:

We should consider decoupling the CLI from the main library. Having a cli to explore the data is nice, however the general use case for distroinfo is a library. Moving the CLI to another repo may make sense, it would decouple things as well as keep the dependencies on the library itself smaller.

Patch-set: 1
Label: Workflow=0