Use python3 in build-python-release
Currently, the job upload-pypi is failing with distroinfo because it
tries to use `python` executable while since centos8 it should use
`python3`. This patch is setting the variable `release_python` which is
used in build-python-release role which is used under the hood [1].

Also, this patch is skipping test test_find_element_in_sub_dict_list
when running in python2 as it fails randomly. Until we can fix the
behavior of find_element in python2, let's skip this test.

[1] https://opendev.org/zuul/zuul-jobs/src/branch/master/roles/build-python-release/defaults/main.yaml

Change-Id: I45f92517a319e31a2e6c84161a364d6ad01745c8